requirements.txt
# Life Fractal Intelligence - Production Dependencies
# ═══════════════════════════════════════════════════════════════════════════════════════

# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Data Processing
numpy==1.24.3
Pillow==10.1.0

# Machine Learning (optional but recommended)
scikit-learn==1.3.2

# GPU Acceleration (optional - only if you have NVIDIA GPU)
# torch==2.1.0+cu118 --index-url https://download.pytorch.org/whl/cu118
# cupy-cuda11x==12.3.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Security
python-dotenv==1.0.0
cryptography==41.0.7

# Monitoring (optional)
# sentry-sdk[flask]==1.39.1

# Payment Integration
# stripe==7.7.0

# Database (if scaling beyond JSON files)
# flask-sqlalchemy==3.1.1
# psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-flask==1.3.0
